requirements.txt
Flask==0.11
kvstore
configuration-registry
python-keyczar
